move workers test to unit tests
