+++ cli/js/tests/unit_tests.ts
@@ -62,5 +62,6 @@ import "./url_test.ts";
 import "./url_search_params_test.ts";
 import "./utime_test.ts";
 import "./write_file_test.ts";
+import "./workers_test.ts";
 import "./performance_test.ts";
 import "./version_test.ts";
